+++ spec/tdc/tdc_version_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe Tdc do
   it "has the expected version number" do
-    expect(Tdc::VERSION).to eq("0.3.7")
+    expect(Tdc::VERSION).to eq("0.3.8")
   end
 end